Revert "Partial rewrite"
