+++ api/tenor.py
@@ -1,5 +1,4 @@
 from aiohttp_requests import requests
-from typing import Optional
 import json
 import asyncio
 import os
@@ -7,15 +6,11 @@ import os
 APIKEY = os.getenv('APIKEY')
 
 
-def _simplify(
-    url: str
-) -> str:
+def _simplify(url):
     return url.replace('http://', '').replace('https://', '')
 
 
-def get_gif_id(
-    url: str
-) -> str:
+def get_gif_id(url: str):
     base_url = 'tenor.com/view/'
     url = _simplify(url.casefold())
     if not url.startswith(base_url):
@@ -25,9 +20,7 @@ def get_gif_id(
     return gif_id
 
 
-async def get_gif_url(
-    gifid: str
-) -> Optional[str]:
+async def get_gif_url(gifid: str):
     r = await requests.get(
         f"https://api.tenor.com/v1/gifs?ids={gifid}&key={APIKEY}"
     )
