+++ cogs/stats.py
@@ -4,6 +4,7 @@ from bot_config import OWNER_ID
 import statcord
 import os
 import dbl
+from pprint import pprint
 
 
 STATCORD_TOKEN = os.getenv("STATCORD_TOKEN")
