+++ cogs/utility.py
@@ -3,9 +3,12 @@ import functions
 import bot_config
 import checks
 from discord.ext.commands import BucketType
+from events import retotal
 from discord.ext import commands
+from events import starboard_events
 from disputils import BotEmbedPaginator
-from cogs.starboard import handle_starboards
+from events import retotal
+from events.starboard_events import handle_starboards
 
 
 async def scan_recount(bot, channel, messages: int, start_date=None):
@@ -14,8 +17,8 @@ async def scan_recount(bot, channel, messages: int, start_date=None):
         start_date = guild.me.created_at
 
     async for m in channel.history(limit=messages):
-        if await functions.needs_recount(bot, m):
-            await functions.recount_reactions(bot, m)
+        if await retotal.needs_recount(bot, m):
+            await retotal.recount_reactions(bot, m)
 
 
 async def handle_trashing(db, bot, ctx, _message_id, trash: bool):
@@ -51,7 +54,7 @@ async def handle_trashing(db, bot, ctx, _message_id, trash: bool):
         message = None
 
     if status is True:
-        await handle_starboards(
+        await starboard_events.handle_starboards(
             db, bot, message_id, channel, message,
             ctx.guild
         )
@@ -119,7 +122,7 @@ async def handle_forcing(
         "Message unforced."
     )
 
-    await handle_starboards(
+    await starboard_events.handle_starboards(
         bot.db, bot, message.id, message.channel, message,
         ctx.guild
     )
@@ -240,7 +243,7 @@ class Utility(commands.Cog):
         except Exception:
             message_obj = None
 
-        await handle_starboards(
+        await starboard_events.handle_starboards(
             self.bot.db, self.bot, message_id, channel,
             message_obj, ctx.guild
         )
@@ -286,7 +289,7 @@ class Utility(commands.Cog):
         except Exception:
             message_obj = None
 
-        await handle_starboards(
+        await starboard_events.handle_starboards(
             self.bot.db, self.bot, mid, channel,
             message_obj, ctx.guild
         )
@@ -380,8 +383,6 @@ class Utility(commands.Cog):
         get_starboard_message = \
             """SELECT * FROM messages WHERE is_orig=False
             AND orig_message_id=$1"""
-
-        sql_sb_messages = []
 
         async with self.db.lock:
             conn = await self.bot.db.connect()
@@ -525,7 +526,7 @@ class Utility(commands.Cog):
         )
 
         async with ctx.typing():
-            await functions.recount_reactions(self.bot, message)
+            await retotal.recount_reactions(self.bot, message)
             await handle_starboards(
                 self.bot.db, self.bot, message.id,
                 message.channel, message, ctx.guild
