+++ database/sql/placeholder.py
@@ -1,0 +1,2 @@
+# Ignore this file. It's just here so that the folder
+# "sql" isn't ignored by github
