+++ functions.py
@@ -2,135 +2,12 @@ from discord import utils
 from discord.ext import commands
 from typing import Tuple, Union, Iterable
 from paginators import disputils
-from api import tenor
-from cogs import starboard
-from itertools import compress
 import emoji
 import bot_config
 import discord
 import functions
 import datetime
 import errors
-
-
-async def needs_recount(bot, message):
-    get_reactions = \
-        """SELECT * FROM reactions WHERE message_id=$1"""
-
-    if message is None:
-        return False
-
-    total = 0
-    reactions = [
-        str(r.emoji.id) if r.custom_emoji else str(r.emoji)
-        for r in message.reactions
-    ]
-    reaction_mask = await functions.is_starboard_emoji(
-        bot.db, message.guild.id, reactions, multiple=True
-    )
-    for r in compress(message.reactions, reaction_mask):
-        total += r.count
-
-    if total == 0:  # Don't recount if the message doesn't have reactions
-        return False
-
-    async with bot.db.lock:
-        conn = bot.db.conn
-        async with conn.transaction():
-            reactions = await conn.fetch(
-                get_reactions, message.id
-            )
-            sql_total = len(reactions)
-
-    if sql_total < 0.5*total and total-sql_total > 2:
-        # recount if the bot has logged less than 10% of the reactions
-        return True
-    return False
-
-
-async def recount_reactions(bot, message):
-    check_reaction = \
-        """SELECT * FROM reactions WHERE
-        message_id=$1 AND name=$2 AND user_id=$3"""
-    check_message = \
-        """SELECT * FROM messages
-        WHERE id=$1"""
-
-    # hard to explain why, but I also remove the message
-    # from the cache when recounting the stars on it
-    await bot.db.cache.remove(message.id, message.guild.id)
-    message = await functions.fetch(bot, message.id, message.channel)
-    if message is None:
-        return
-
-    # [{'user_id': user_id, 'name': name}, ...]
-    # other values can be determined from the message object
-    to_add = []
-
-    for reaction in message.reactions:
-        if reaction.custom_emoji:
-            name = str(reaction.emoji.id)
-        else:
-            name = str(reaction.emoji)
-
-        if not await functions.is_starboard_emoji(
-            bot.db, message.guild.id, name
-        ):
-            continue
-
-        async for user in reaction.users():
-            if user is None:
-                continue
-            elif user.bot:
-                continue
-            to_add.append({
-                'user': user, 'name': name
-            })
-
-    conn = bot.db.conn
-
-    async with bot.db.lock:
-        async with conn.transaction():
-            sql_m = await conn.fetchrow(
-                check_message, message.id
-            )
-            if sql_m and sql_m['is_orig'] is False:
-                print("No")
-                return
-            elif sql_m is None:
-                await bot.db.q.create_message.fetch(
-                    message.id, message.guild.id,
-                    message.author.id, None,
-                    message.channel.id, True,
-                    message.channel.is_nsfw()
-                )
-
-    for r in to_add:
-        await functions.check_or_create_existence(
-            bot,
-            guild_id=message.guild.id,
-            user=r['user'], do_member=True
-        )
-
-        async with bot.db.lock:
-            async with conn.transaction():
-                sql_r = await conn.fetchrow(
-                    check_reaction, message.id, r['name'],
-                    r['user'].id
-                )
-                if sql_r is not None:
-                    continue
-
-                await bot.db.q.create_reaction.fetch(
-                    message.guild.id, r['user'].id,
-                    message.id, r['name']
-                )
-
-    await starboard.handle_starboards(
-        bot.db, bot, message.id, message.channel, message,
-        message.guild
-    )
-
 
 async def is_starboard_emoji(db, guild_id, emoji, multiple=False):
     if not multiple:
@@ -154,182 +31,6 @@ async def is_starboard_emoji(db, guild_id, emoji, multiple=False):
         return str(emoji) in all_emojis
     else:
         return [emo in all_emojis for emo in emoji]
-
-
-async def get_embed_from_message(message):
-    nsfw = message.channel.is_nsfw()
-    embed = discord.Embed(
-        title="NSFW" if nsfw else discord.Embed.Empty, colour=bot_config.COLOR
-    )
-    embed.set_author(
-        name=str(message.author), icon_url=message.author.avatar_url
-    )
-    embed_text = ''
-    msg_attachments = message.attachments
-    urls = []
-
-    for attachment in msg_attachments:
-        urls.append({
-            'name': attachment.filename, 'display_url': attachment.url,
-            'url': attachment.url, 'type': 'upload'
-        })
-
-    e = discord.embeds._EmptyEmbed
-
-    for msg_embed in message.embeds:
-        if msg_embed.type == 'rich':
-            fields = [
-                (
-                    f"\n**{x.name if type(x.name) != e else ''}**\n",
-                    f"{x.value if type(x.value) != e else ''}\n"
-                )
-                for x in msg_embed.fields
-            ]
-            embed_text += f"__**{msg_embed.title}**__\n"\
-                if type(msg_embed.title) != e else ''
-            embed_text += f"{msg_embed.description}\n"\
-                if type(msg_embed.description) != e else ''
-
-            for name, value in fields:
-                embed_text += name + value
-            if msg_embed.footer.text is not embed.Empty:
-                embed_text += '\n' + str(msg_embed.footer.text) + '\n'
-            if msg_embed.image.url is not embed.Empty:
-                urls.append({
-                    'name': 'Embed Image',
-                    'url': msg_embed.image.url,
-                    'display_url': msg_embed.image.url
-                })
-            if msg_embed.thumbnail.url is not embed.Empty:
-                urls.append({
-                    'name': 'Embed Thumbnail',
-                    'url': msg_embed.thumbnail.url,
-                    'display_url': msg_embed.thumbnail.url
-                })
-        elif msg_embed.type == 'image':
-            if msg_embed.url != discord.Embed.Empty:
-                urls.append({
-                    'name': 'Image', 'display_url': msg_embed.thumbnail.url,
-                    'url': msg_embed.url, 'type': 'image'
-                })
-        elif msg_embed.type == 'gifv':
-            gifid = tenor.get_gif_id(msg_embed.url)
-            if gifid is None:
-                display_url = msg_embed.thumbnail.url
-            else:
-                display_url = await tenor.get_gif_url(gifid)
-            if msg_embed.url != discord.Embed.Empty:
-                urls.append({
-                    'name': 'GIF', 'display_url': display_url,
-                    'url': msg_embed.url, 'type': 'gif'
-                })
-        elif msg_embed.type == 'video':
-            if msg_embed.url != discord.Embed.Empty:
-                urls.append({
-                    'name': 'Video', 'display_url': msg_embed.thumbnail.url,
-                    'url': msg_embed.url, 'type': 'video'
-                })
-
-    value_string = f"{message.system_content}\n{embed_text}"
-    context_string = f"\n[**Jump to Message**]({message.jump_url})"
-    if len(value_string) > 2048:
-        clip_msg = "... *message clipped*"
-        to_clip = len(value_string+clip_msg)-2048
-        full_string = value_string[0:-1*to_clip] + clip_msg
-    else:
-        full_string = value_string
-    embed.description = full_string
-
-    embed.add_field(name="Original", value=context_string)
-
-    if len(urls) > 0:
-        url_string = ''
-        current = 0
-        for item in urls:
-            url_string += f"[**{item['name']}**]({item['url']})\n"
-            if current == 0:
-                embed.set_image(url=item['display_url'])
-                current += 1
-            elif current == 1:
-                embed.set_thumbnail(url=item['display_url'])
-                current += 1
-        embed.add_field(name='Attachments', value=url_string, inline=False)
-
-    embed.set_footer(text=f"ID: {message.id}")
-    embed.timestamp = message.created_at
-
-    return embed
-
-
-async def calculate_points(conn, sql_message, sql_starboard, bot, guild):
-    get_reactions = \
-        """SELECT * FROM reactions WHERE message_id=$1"""
-    get_user = \
-        """SELECT * FROM users WHERE id=$1"""
-    get_sbemojis = \
-        """SELECT * FROM sbemojis WHERE starboard_id=$1"""
-    update_message = \
-        """UPDATE messages
-        SET points=$1
-        WHERE orig_message_id=$2
-        AND channel_id=$3"""
-
-    message_id = int(sql_message['id'])
-    self_star = sql_starboard['self_star']
-
-    async with bot.db.lock:
-        async with conn.transaction():
-            emojis = await conn.fetch(get_sbemojis, sql_starboard['id'])
-            all_reactions = await conn.fetch(get_reactions, message_id)
-
-    used_users = set()
-
-    total_points = 0
-    for emoji_obj in emojis:
-        emoji_id = int(emoji_obj['d_id']) if emoji_obj['d_id'] is not None\
-            else None
-        emoji_name = None if emoji_id is not None else emoji_obj['name']
-        reactions = [
-            r for r in all_reactions if r['name']
-            in [str(emoji_id), emoji_name]
-        ]
-        for sql_reaction in reactions:
-            user_id = sql_reaction['user_id']
-            if user_id in used_users:
-                continue
-            used_users.add(user_id)
-            if user_id == sql_message['user_id'] and self_star is False:
-                continue
-
-            async with bot.db.lock:
-                async with conn.transaction():
-                    sql_user = await conn.fetchrow(get_user, user_id)
-
-            if sql_user['is_bot'] is True:
-                continue
-
-            member_list = await functions.get_members(
-                [int(sql_user['id'])], guild
-            )
-            try:
-                member = member_list[0]
-                if member and await functions.is_user_blacklisted(
-                    bot, member, int(sql_starboard['id'])
-                ):
-                    continue
-            except IndexError:
-                pass
-
-            total_points += 1
-
-    async with bot.db.lock:
-        async with conn.transaction():
-            await conn.execute(
-                update_message, total_points,
-                message_id, int(sql_starboard['id'])
-            )
-
-    return total_points, emojis
 
 
 async def get_members(user_ids: Iterable[int], guild: discord.Guild):
